76284: non site admin edit collection authorization group REST feedback
- Merge AuthorizeSolrService into AuthorizeService
- PreAuthorize findAdminAuthorized methods
