+++ dspace-api/src/main/java/org/dspace/authorize/factory/AuthorizeServiceFactory.java
@@ -8,7 +8,6 @@
 package org.dspace.authorize.factory;
 
 import org.dspace.authorize.service.AuthorizeService;
-import org.dspace.authorize.service.AuthorizeSolrService;
 import org.dspace.authorize.service.ResourcePolicyService;
 import org.dspace.services.factory.DSpaceServicesFactory;
 
@@ -23,8 +22,6 @@ public abstract class AuthorizeServiceFactory {
     public abstract AuthorizeService getAuthorizeService();
 
     public abstract ResourcePolicyService getResourcePolicyService();
-
-    public abstract AuthorizeSolrService getAuthorizeSolrService();
 
     public static AuthorizeServiceFactory getInstance() {
         return DSpaceServicesFactory.getInstance().getServiceManager()
