+++ dspace-api/src/main/java/org/dspace/authorize/factory/AuthorizeServiceFactoryImpl.java
@@ -8,7 +8,6 @@
 package org.dspace.authorize.factory;
 
 import org.dspace.authorize.service.AuthorizeService;
-import org.dspace.authorize.service.AuthorizeSolrService;
 import org.dspace.authorize.service.ResourcePolicyService;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -24,8 +23,6 @@ public class AuthorizeServiceFactoryImpl extends AuthorizeServiceFactory {
     private AuthorizeService authorizeService;
     @Autowired(required = true)
     private ResourcePolicyService resourcePolicyService;
-    @Autowired(required = true)
-    private AuthorizeSolrService authorizeSolrService;
 
     @Override
     public AuthorizeService getAuthorizeService() {
@@ -35,10 +32,5 @@ public class AuthorizeServiceFactoryImpl extends AuthorizeServiceFactory {
     @Override
     public ResourcePolicyService getResourcePolicyService() {
         return resourcePolicyService;
-    }
-
-    @Override
-    public AuthorizeSolrService getAuthorizeSolrService() {
-        return authorizeSolrService;
     }
 }
