+++ dspace-server-webapp/src/main/java/org/dspace/app/rest/authorization/impl/CommunityAdminFeature.java
@@ -14,7 +14,6 @@ import org.dspace.app.rest.authorization.AuthorizationFeatureDocumentation;
 import org.dspace.app.rest.model.BaseObjectRest;
 import org.dspace.app.rest.model.SiteRest;
 import org.dspace.authorize.service.AuthorizeService;
-import org.dspace.authorize.service.AuthorizeSolrService;
 import org.dspace.core.Context;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -34,8 +33,6 @@ public class CommunityAdminFeature implements AuthorizationFeature {
 
     @Autowired
     private AuthorizeService authorizeService;
-    @Autowired
-    private AuthorizeSolrService authorizeSolrService;
 
     @Override
     public boolean isAuthorized(Context context, BaseObjectRest object) throws SQLException {
@@ -44,7 +41,7 @@ public class CommunityAdminFeature implements AuthorizationFeature {
                 return true;
             }
 
-            return authorizeSolrService.isCommunityAdmin(context);
+            return authorizeService.isCommunityAdminSolr(context);
         }
         return false;
     }
