+++ dspace-server-webapp/src/main/java/org/dspace/app/rest/repository/CommunityRestRepository.java
@@ -34,7 +34,6 @@ import org.dspace.app.rest.model.patch.Patch;
 import org.dspace.app.rest.utils.CommunityRestEqualityUtils;
 import org.dspace.authorize.AuthorizeException;
 import org.dspace.authorize.service.AuthorizeService;
-import org.dspace.authorize.service.AuthorizeSolrService;
 import org.dspace.content.Bitstream;
 import org.dspace.content.Community;
 import org.dspace.content.service.BitstreamService;
@@ -84,9 +83,6 @@ public class CommunityRestRepository extends DSpaceObjectRestRepository<Communit
     AuthorizeService authorizeService;
 
     private CommunityService cs;
-
-    @Autowired
-    private AuthorizeSolrService authorizeSolrService;
 
     public CommunityRestRepository(CommunityService dsoService) {
         super(dsoService);
@@ -208,15 +204,16 @@ public class CommunityRestRepository extends DSpaceObjectRestRepository<Communit
         }
     }
 
+    @PreAuthorize("hasAuthority('AUTHENTICATED')")
     @SearchRestMethod(name = "findAdminAuthorized")
     public Page<CommunityRest> findAdminAuthorized (
         Pageable pageable, @Parameter(value = "query") String query) {
         try {
             Context context = obtainContext();
-            List<Community> communities = authorizeSolrService.findAdminAuthorizedCommunity(context, query,
+            List<Community> communities = authorizeService.findAdminAuthorizedCommunity(context, query,
                 Math.toIntExact(pageable.getOffset()),
                 Math.toIntExact(pageable.getPageSize()));
-            Long tot = authorizeSolrService.countAdminAuthorizedCommunity(context, query);
+            Long tot = authorizeService.countAdminAuthorizedCommunity(context, query);
             return converter.toRestPage(communities, pageable, tot , utils.obtainProjection());
         } catch (SearchServiceException | SQLException e) {
             throw new RuntimeException(e.getMessage(), e);
